Filter unknown algorithm
This commit modifies the way the pub_key_cred_params is parsed to silently drop unknown algorithms.
This ensures that the deserialization does not fail if the host sends more than 12 algorithms,i
since we know how many algorithms are supported and will therefore be parsed.

Regarding the spec:
  This member contains information about the desired properties of the credential to be created.
  The sequence is ordered from most preferred to least preferred.
  The client makes a best-effort to create the most preferred credential that it can.

Since the behaviour is best-effort, it makes sense to drop unsupported algorithms at serialization.

We could even consider only keeping the first algorithm that it supported since it is the one that will be used.